refactor: move json_converters to shared package
