--- catalyst_voices/packages/internal/catalyst_voices_models/lib/src/json_converters.dart
@@ -1,23 +1,0 @@
-import 'package:catalyst_cardano_serialization/catalyst_cardano_serialization.dart';
-import 'package:json_annotation/json_annotation.dart';
-
-final class CoinConverter implements JsonConverter<Coin, int> {
-  const CoinConverter();
-
-  @override
-  Coin fromJson(int json) => Coin(json);
-
-  @override
-  int toJson(Coin object) => object.value;
-}
-
-final class ShelleyAddressConverter
-    implements JsonConverter<ShelleyAddress, String> {
-  const ShelleyAddressConverter();
-
-  @override
-  ShelleyAddress fromJson(String json) => ShelleyAddress.fromBech32(json);
-
-  @override
-  String toJson(ShelleyAddress object) => object.toBech32();
-}
