+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/wallet/wallet_info.dart
@@ -1,7 +1,7 @@
 import 'package:catalyst_cardano/catalyst_cardano.dart';
 import 'package:catalyst_cardano_serialization/catalyst_cardano_serialization.dart';
-import 'package:catalyst_voices_models/src/json_converters.dart';
 import 'package:catalyst_voices_models/src/wallet/wallet_metadata.dart';
+import 'package:catalyst_voices_shared/catalyst_voices_shared.dart';
 import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
